// src/generated/api/ledger/schema.ts
export interface components {
    schemas: {
        AccountBalance: {
            account_id: string;
            asset_code: string;
            /** Format: int64 */
            balance: number;
        };
        AccountDetail: {
            account_type: string;
            asset_code: string;
            /** Format: date-time */
            created_at: string;
            id: string;
            owner_id: string;
            owner_type: string;
            status: string;
        };
        EntryDetail: {
            account_id: string;
            /** Format: int64 */
            amount: number;
            asset_code: string;
            direction: string;
            id: string;
        };
        OpenAccountRequest: {
            account_type: string;
            asset_code: string;
            owner_id: string;
            owner_type: string;
        };
        PostEntryRequest: {
            account_id: string;
            /** Format: int64 */
            amount: number;
            direction: string;
        };
        PostTransactionRequest: {
            description: string;
            entries: components["schemas"]["PostEntryRequest"][];
            idempotency_key: string;
            market_symbol?: string;
            source_module: string;
            source_operation_id: string;
            source_operation_type: string;
        };
        ProblemFieldError: {
            code: string;
            message: string;
        };
        ProblemResponse: {
            code: string;
            errors?: components["schemas"]["ProblemFieldError"][];
            message: string;
        };
        ReverseTransactionRequest: {
            description: string;
            idempotency_key: string;
            source_module: string;
            source_operation_id: string;
            source_operation_type: string;
        };
        TransactionDetail: {
            description: string;
            entries: components["schemas"]["EntryDetail"][];
            id: string;
            idempotency_key: string;
            market_symbol?: string;
            /** Format: date-time */
            posted_at: string;
            reverses_transaction_id?: string;
            source_module: string;
            source_operation_id: string;
            source_operation_type: string;
            transaction_type: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}